Move code from responders.run to testrunner package, removing package dependency
