+++ src/fitnesse/responders/run/SuiteResponder.java
@@ -4,6 +4,7 @@ package fitnesse.responders.run;
 
 import java.io.IOException;
 
+import fitnesse.http.Request;
 import fitnesse.reporting.BaseFormatter;
 import fitnesse.reporting.CachingSuiteXmlFormatter;
 import fitnesse.testrunner.CompositeFormatter;
@@ -11,9 +12,16 @@ import fitnesse.reporting.PageHistoryFormatter;
 import fitnesse.reporting.history.SuiteHistoryFormatter;
 import fitnesse.reporting.SuiteHtmlFormatter;
 import fitnesse.testrunner.MultipleTestsRunner;
+import fitnesse.testrunner.SuiteContentsFinder;
+import fitnesse.testrunner.SuiteFilter;
 import fitnesse.testsystems.TestSystemListener;
 
 public class SuiteResponder extends TestResponder {
+  private static final String NOT_FILTER_ARG = "excludeSuiteFilter";
+  private static final String AND_FILTER_ARG = "runTestsMatchingAllTags";
+  private static final String OR_FILTER_ARG_1 = "runTestsMatchingAnyTag";
+  private static final String OR_FILTER_ARG_2 = "suiteFilter";
+
   private boolean includeHtml;
 
   @Override
@@ -60,9 +68,56 @@ public class SuiteResponder extends TestResponder {
 
   @Override
   protected void performExecution() throws IOException, InterruptedException {
-    SuiteFilter filter = new SuiteFilter(request, page.getPageCrawler().getFullPath().toString());
+    SuiteFilter filter = createSuiteFilter(request, page.getPageCrawler().getFullPath().toString());
     SuiteContentsFinder suiteTestFinder = new SuiteContentsFinder(page, filter, root);
     MultipleTestsRunner runner = newMultipleTestsRunner(suiteTestFinder.getAllPagesToRunForThisSuite());
     runner.executeTestPages();
   }
+
+  public static SuiteFilter createSuiteFilter(Request request, String suitePath) {
+    return new SuiteFilter(getOrTagFilter(request),
+            getNotSuiteFilter(request),
+            getAndTagFilters(request),
+            getSuiteFirstTest(request, suitePath));
+  }
+
+  private static String getOrTagFilter(Request request) {
+    return request != null ? getOrFilterString(request) : null;
+  }
+
+  private static String getOrFilterString(Request request) {
+    //request already confirmed not-null
+    String orFilterString = null;
+    if(request.getInput(OR_FILTER_ARG_1) != null){
+      orFilterString = (String) request.getInput(OR_FILTER_ARG_1);
+    } else {
+      orFilterString = (String) request.getInput(OR_FILTER_ARG_2);
+    }
+    return orFilterString;
+  }
+
+  private static String getNotSuiteFilter(Request request) {
+    return request != null ? (String) request.getInput(NOT_FILTER_ARG) : null;
+  }
+
+  private static String getAndTagFilters(Request request) {
+    return request != null ? (String) request.getInput(AND_FILTER_ARG) : null;
+  }
+
+
+  private static String getSuiteFirstTest(Request request, String suiteName) {
+    String startTest = null;
+    if (request != null) {
+      startTest = (String) request.getInput("firstTest");
+    }
+
+    if (startTest != null) {
+      if (startTest.indexOf(suiteName) != 0) {
+        startTest = suiteName + "." + startTest;
+      }
+    }
+
+    return startTest;
+  }
+
 }
