+++ src/fitnesse/testrunner/SuiteFilter.java
@@ -1,25 +1,19 @@
 /**
  * Copyright AdScale GmbH, Germany, 2009
  */
-package fitnesse.responders.run;
-
-import util.StringUtil;
-
-import fitnesse.http.Request;
-import fitnesse.wiki.PageCrawler;
-import fitnesse.wiki.PageData;
-import fitnesse.wiki.WikiPage;
-import fitnesse.wiki.WikiPagePath;
+package fitnesse.testrunner;
 
 import java.util.Arrays;
 import java.util.LinkedList;
 import java.util.List;
 
+import fitnesse.wiki.PageCrawler;
+import fitnesse.wiki.PageData;
+import fitnesse.wiki.WikiPage;
+import fitnesse.wiki.WikiPagePath;
+import util.StringUtil;
+
 public class SuiteFilter {
-  private static final String NOT_FILTER_ARG = "excludeSuiteFilter";
-  private static final String AND_FILTER_ARG = "runTestsMatchingAllTags";
-  private static final String OR_FILTER_ARG_1 = "runTestsMatchingAnyTag";
-  private static final String OR_FILTER_ARG_2 = "suiteFilter";
   final private SuiteTagMatcher notMatchTags;
   final private SuiteTagMatcher matchTags;
   final private boolean andStrategy;
@@ -33,7 +27,7 @@ public class SuiteFilter {
   
   public static final SuiteFilter MATCH_ALL = new SuiteFilter(null, null, null, null);
 
-  SuiteFilter(String orTags, String mustNotMatchTags, String andTags, String startWithTest) {
+  public SuiteFilter(String orTags, String mustNotMatchTags, String andTags, String startWithTest) {
     this.startWithTest = (!"".equals(startWithTest)) ? startWithTest : null;
     if(andTags != null){
       matchTags = new SuiteTagMatcher(andTags, true);
@@ -45,13 +39,6 @@ public class SuiteFilter {
     notMatchTags = new SuiteTagMatcher(mustNotMatchTags, false);
   }
 
-  public SuiteFilter(Request request, String suitePath) {
-    this(getOrTagFilter(request), 
-        getNotSuiteFilter(request),
-        getAndTagFilters(request),
-        getSuiteFirstTest(request, suitePath));
-  }
-
   public SuiteFilter(String suiteFilter, String excludeSuiteFilter) {
     matchTags = new SuiteTagMatcher(suiteFilter, true);
     notMatchTags = new SuiteTagMatcher(excludeSuiteFilter, false);
@@ -59,45 +46,6 @@ public class SuiteFilter {
     startWithTest = null;
   }
 
-  private static String getOrTagFilter(Request request) {
-    return request != null ? getOrFilterString(request) : null;
-  }
-
-  private static String getOrFilterString(Request request) {
-    //request already confirmed not-null
-    String orFilterString = null;
-    if(request.getInput(OR_FILTER_ARG_1) != null){
-      orFilterString = (String) request.getInput(OR_FILTER_ARG_1);
-    } else {
-      orFilterString = (String) request.getInput(OR_FILTER_ARG_2);
-    }
-    return orFilterString;
-  }
-
-  private static String getNotSuiteFilter(Request request) {
-    return request != null ? (String) request.getInput(NOT_FILTER_ARG) : null;
-  }
-  
-  private static String getAndTagFilters(Request request) {
-    return request != null ? (String) request.getInput(AND_FILTER_ARG) : null;
-  }
-
-
-  private static String getSuiteFirstTest(Request request, String suiteName) {
-    String startTest = null;
-    if (request != null) {
-      startTest = (String) request.getInput("firstTest");
-    }
-
-    if (startTest != null) {
-      if (startTest.indexOf(suiteName) != 0) {
-        startTest = suiteName + "." + startTest;
-      }
-    }
-
-    return startTest;
-  }
-  
   public boolean isMatchingTest(WikiPage testPage) {
     PageData data = testPage.getData();
     boolean pruned = data.hasAttribute(PageData.PropertyPRUNE);
