+++ src/fitnesse/testrunner/TestsRunnerListener.java
@@ -1,7 +1,5 @@
 package fitnesse.testrunner;
 
-import fitnesse.reporting.CompositeExecutionLog;
-
 public interface TestsRunnerListener {
 
   void setTrackingId(String stopResponderId);
